pip-requirements-test.txt
tox>=2.0.0
pylint>=2.6.0
django-override-settings>=1.2
twine>=1.13.0
yapf==0.28.0
pre-commit==1.14.4